:recycle: refactor: Rework contexts
Given the necessity to turn on and off loading, this was remade into a
component instead, so there is no need to keep a separate state
management just for this visual element.
The remaining contexts were mainly renamed so they would be less
convoluted. NotificationContext now only deals with errors, so it became
ErrorContext. UserStatusContext became UserContext because the "status"
word was unnecessary.